pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-generativeai>=0.8.5",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-qdrant>=0.2.0",
    "langchain-text-splitters>=0.3.8",
    "langchain[google-genai]>=0.3.26",
    "ollama>=0.5.1",
    "pypdf>=5.7.0",
]
